added proper isOwner check on collection, removed extra argument in index.tsx

+++ www/pages/collections/[cid].tsx
@@ -1,13 +1,13 @@
 import type { GetServerSideProps, GetServerSidePropsContext, NextPage } from "next";
 import Image from "next/image";
 
-import { IUser, ISyllabus } from "types";
+import { IUser, ISyllabus, ICollection } from "types";
 import BreadcrumbsBar from "components/commons/BreadcrumbsBar";
 
 import { getSyllabusCards } from "components/utils/getSyllabusCards";
 import NotFound from "components/commons/NotFound";
 import { signOut, useSession } from "next-auth/react";
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import Router from "next/router";
 import Link from "next/link";
 import { getToken } from "next-auth/jwt";
@@ -17,10 +17,11 @@ import cancelIcon from '../../public/icons/close-line.svg'
 import checkIcon from '../../public/icons/check-line.svg'
 import deleteIcon from '../../public/icons/delete-bin-line.svg'
 import removeIcon from '../../public/icons/subtract-line.svg'
+import { Session } from "next-auth";
 
 export const getServerSideProps: GetServerSideProps = async (context: GetServerSidePropsContext) => {
   const collectionId = context.params!.cid;
-  
+
   const t = await getToken({ req: context.req, secret: process.env.NEXTAUTH_SECRET })
   const token = t ? (t.user as { _id: string, token: string }).token : '';
   const url = new URL(`collections/${collectionId}`, process.env.NEXT_PUBLIC_API_URL);
@@ -33,7 +34,7 @@ export const getServerSideProps: GetServerSideProps = async (context: GetServerS
   if (res.ok) {
     const collectionInfo = await res.json();
     return {
-      props: collectionInfo,
+      props: { collectionInfo: collectionInfo },
     };
   } else {
     return {
@@ -42,25 +43,37 @@ export const getServerSideProps: GetServerSideProps = async (context: GetServerS
   }
 };
 
-interface ICollectionProps {
-  uuid: string;
-  status: string;
-  name: string;
-  description: string;
-  syllabi: ISyllabus[];
-  user_uuid: string;
-  user: IUser;
+interface ICollectionPageProps {
+  collectionInfo: ICollection
 }
 
-const Collection: NextPage<ICollectionProps> = (props) => {
+const Collection: NextPage<ICollectionPageProps> = ({ collectionInfo }) => {
 
   const { data: session, status } = useSession();
-
   const [isEditing, setIsEditing] = useState(false);
-  const [name, setName] = useState(props.name || '')
-  const [tmp, setTmp] = useState(props.name || '')
+  const [name, setName] = useState('')
+  const [tmp, setTmp] = useState('')
   const [log, setLog] = useState('')
-  const url = new URL(`/collections/${props.uuid}`, process.env.NEXT_PUBLIC_API_URL)
+  const [isOwner, setIsOwner] = useState(false)
+  const [endpoint, setEndpoint] = useState<URL>()
+
+  const checkIfOwner = (_session: Session, _uuid: string) => {
+    if (_session.user != null) {
+      return _session.user._id === _uuid;
+    }
+    return false
+  };
+
+  useEffect(() => {
+    if (!collectionInfo || !session) return
+    const o = checkIfOwner(session, collectionInfo.user_uuid)
+    setName(collectionInfo.name)
+    setTmp(collectionInfo.name)
+    setEndpoint(new URL(`/collections/${collectionInfo.uuid}`, process.env.NEXT_PUBLIC_API_URL))
+
+    setIsOwner(o)
+  }, [session, collectionInfo])
+
   const confirmMsg = `Do you really want to delete the collection ${name}? This action cannot be undone.`;
   const confirmRemoveMsg = `Do you really want to remove this syllabus from ${name}? This action cannot be undone.`
 
@@ -73,27 +86,20 @@ const Collection: NextPage<ICollectionProps> = (props) => {
     tags_exclude: [],
   }
 
-  if (Object.keys(props).length === 0) {
-    return (
-      <NotFound />
-    )
-  }
-
-  const checkIfAdmin = () => {
-    if (session != null && session.user != null) {
-      return session.user._id === props.user_uuid;
-    }
-    return false
-  };
-
   const submitEdit = () => {
+    if (!endpoint) return;
+
     const h = new Headers();
     h.append("Authorization", `Bearer ${session?.user.token}`);
 
     let b = new FormData()
+    if (tmp.length < 1) {
+      setLog("Name must be at least 1 character long")
+      return;
+    }
     b.append("name", tmp)
 
-    fetch(url, {
+    fetch(endpoint, {
       method: 'PATCH',
       headers: h,
       body: b
@@ -118,13 +124,14 @@ const Collection: NextPage<ICollectionProps> = (props) => {
   }
 
   const submitDelete = () => {
+    if (!endpoint) return;
     if (!window.confirm(confirmMsg))
       return
 
     const h = new Headers();
     h.append("Authorization", `Bearer ${session?.user.token}`);
 
-    fetch(url, {
+    fetch(endpoint, {
       method: 'DELETE',
       headers: h,
     })
@@ -153,7 +160,7 @@ const Collection: NextPage<ICollectionProps> = (props) => {
   const handleRemove = (e: React.BaseSyntheticEvent) => {
     e.preventDefault()
     const t = e.target;
-    const removeUrl = new URL(`/collections/${props.uuid}/syllabi/${t.dataset.syllabusid}`, process.env.NEXT_PUBLIC_API_URL)
+    const removeUrl = new URL(`/collections/${collectionInfo.uuid}/syllabi/${t.dataset.syllabusid}`, process.env.NEXT_PUBLIC_API_URL)
 
     if (!window.confirm(confirmRemoveMsg))
       return
@@ -182,14 +189,19 @@ const Collection: NextPage<ICollectionProps> = (props) => {
       })
   }
 
+  if (!collectionInfo)
+    return (
+      <NotFound />
+    )
+
   return (
     <>
       <div className="w-11/12 md:w-10/12 m-auto mt-8">
         <BreadcrumbsBar
-          user={props.user.name}
-          userId={props.user_uuid}
+          user={collectionInfo.user.name}
+          userId={collectionInfo.user_uuid}
           category="collections"
-          pageTitle={props.name}
+          pageTitle={collectionInfo.name}
         />
         <div className="flex flex-col gap-2 my-6">
           {isEditing ?
@@ -204,7 +216,7 @@ const Collection: NextPage<ICollectionProps> = (props) => {
                 </button>
               </div>
             </>
-            : checkIfAdmin() ?
+            : isOwner ?
               <>
                 <h1 className={`${kurintoBook.className} text-3xl`}>{name}</h1>
                 <div className="flex gap-3">
@@ -224,10 +236,10 @@ const Collection: NextPage<ICollectionProps> = (props) => {
 
         </div>
         <div className="gap-3 pb-5">{
-          props.syllabi ?
-            props.syllabi.map(s => {
+          collectionInfo.syllabi ?
+            collectionInfo.syllabi.map(s => {
               return <div key={s.uuid} className="flex gap-1">
-                {getSyllabusCards([s], default_filters, checkIfAdmin())}
+                {getSyllabusCards([s], default_filters)}
                 <div className="border-2 border-gray-900 p-2 rounded-md flex items-center">
                   <button data-syllabusid={s.uuid} onClick={handleRemove}>
                     <Image src={removeIcon} width="24" height="24" alt="Icon to remove an element from the list" />
@@ -237,7 +249,10 @@ const Collection: NextPage<ICollectionProps> = (props) => {
 
             })
             :
-            <>There are no syllabi in this collection. {checkIfAdmin() ? <><Link href="/" className="underline">Browse syllabi</Link> to add them to your collection.</> : <></>}</>
+            <div>
+              <div>There are no syllabi in this collection.</div>
+              {isOwner ? <div className="mt-8"><Link href="/" className="underline">Browse syllabi</Link> to add them to your collection.</div> : <></>}
+            </div>
         }</div>
       </div>
     </>
